set rsync options for merge, removed rsync wrapper
Had to remove the rsync wrapper - there was a strange issue with quoting.
In the main code
rsync $shortOpts $longOpts "$longOptsBackup" [SOURCE] [DEST]
worked. Note the quotes around "$longOptsBackup". They're because
the address might contain spaces. I tried to move the code
for branching rsync re. copying a file vs. copying a folder
into a separate function, since it's now used multiple times.
But when passing $unquotedOpts $quotedOpts to it, and then it calling
rsync $unquotedOpts "$quotedOpts" [SOURCE] [DEST]
a blank $quotedOpts (e.g. if not using backups)
would cause an error. I couldn't find out how to make it work
and even if I could it would be too subtle a distinction to allow
in the code - too hard for a programmer to understand.
So I decided it was better to just duplicate the code.