resolved case UnexpectedlyNotOnUpToDateList into 2
The unusual conditions leading to the $WARNINGUnexpectedlyNotOnUpToDateList case (near
the end of main) can be resolved into two distinct cases: host's mod time is newer, or
removable's mod time is newer. The former implies a fork, I think, while the latter is
actually a fine history.
The message $WARNINGUnexpectedlyNotOnUpToDateList warning, which amounts to saying
"I am confused"
has been replaced with either "it's been forked" or "this is fine" respectively.
HOWEVER, the logic of this is a little tricky, but I can see that if there was a fork
between the rmvbl and anthrhost (see code comments fro explanation of terminology
anthrhst), which the user resolved favouring anthrhst, the program will now NOT
notice if the rmvbl and host have independent changes that the user has not been asked
to choose between.
I think implementing that check would require a big reprogram. So users who've forked
their work - hopefully very few - will have to be careful. Of course, the backup system
will prevent them from losing work.