removed the diffItems check for identical items
Until now, if it determined that the host version and removable drive version should be identical, then it did a diff to check that they were. It has never happened to me that it has thought things were identical but then the diff found otherwise. The diff often takes a long time. So I have commented out the diff step.