added interactive mode dialog for first-time syncs
Interactive mode is supposed to prompt for permission before doing anything,
but previously if doing a first-time sync it would not.
That is to say, if, for example, a file/folder "A" was listed for syncing and
existed on the host but not on the removable drive then it would
copy "A" to the removable drive without asking, even in interactive mode.
There was no risk of loss of data. This only applied to non-interactively
creating a fresh file-folder where none existed before.