added information about unclear error handling
if the pivotal "rsync" command fails, the program basically handles it and nothing should get damaged, but the code is unclear and needs improving.
It's also possible that the unclear code is hiding a possible way that the failed rsync command would result in serious bad behaviour.
Put a note explaining all of this.