Further improved safety of the rm command
Added a check that there are no asterixes (i.e. wildcards) anywhere in the variable (oldBackupName) that's supposed to contain the removed path. (This is probably overkill now.)
Added a diagnostic message for if there's ever a problem. It echoes value of the variable (oldBackupName) containing the invalid value.
This is just me worrying after reading a rm horror story. To date I have never experienced any kind of bug while using the current version of the program, and I use it several times a day.