Further streamlined the rmeoveOldBackups function
Removed the check "oldBackupsExist -eq true" since the case that there are no old backups is already covered in the later code.
Changed the "if $oldBackups is empty" test to exit if it is empty, rather than have the rest of the function inside an "if [it is not empty]" statement.