Removed old commented code
especially the code relating to rsyncFileOrFolderWrapper
which was intended to wrap the branching logic for files/folders
but it was difficult to handle passing it the necessary arguments,
because bash just passes all arguments together in a string,
which then has to be parsed again (i.e. split on the delimeter)
by the function.
This did work, but the code was dreadfully unclear so I dropped it
and went in favour of just duplicated code instead.