Implented summary table for sync/merged items only
Now, at the end of the script, it prints out a table of all the syncs/merges it has done
ONLY WORKS FOR SYNCED/MERGED ITEMS - i.e. does not print anything for items that were SKIPPED
To do this I had to change the while loop over locsList to be redirected into instead of piped into.
This is a big change - it means that changes made to variables (with appropriate scope) in the loop will now also be readable outside the loop