Made merges always count as modifications
After a merge, the top-level of the dest was touched to make it have the same mod time as the source.
This is not correct - a merge is an active modification of the dest, such that the source and dest are NOT the same even after the process is complete.
Touching the top-level of the dest to make its mod time be NOW ensures that it will be recognised as a modification in the correct way.