chooseVersionDialog, auto mode, now uses merge
chooseVersionDialog, in automatic mode, used to synchronise the newer over the older.
However, it's dangerous to automatically pick a version from two forked versions.
Really, the user needs to consider the forks before we know which is better.
I decided that using "merge" instead would prevent lost data, and is the better choice for automatic mode