rejuvenated the tester and added it
I decided that I want to change the way the logic is implemented, the nested "ifs", etc.
and to make some changes to the logic.
Naturally the first step is to write a tester.
I've been using a rough-and-ready tester, but I've polished it up ready for the world to see.
I've also edited the tests to reflect some changes to the logic that I'm considering (so some of them fail currently)