tester output now shows description for each test
using a global variable, that is not readonly (!), edited the output of the tester to display each unit's description, as well as its result.
This should make testing faster.
I realise the $report variable doesn't have to be global, and global mutable variables are bad form, but I'll bend the rules because, after all, this is only a tester!