expanded testing, tweaked messages
expanded the tester to deal with cases where there is no sync record, an invalid sync record, the item is missing from the host, the item is missing from the rmvbl.
Changed the tester's requirements on the messages displayed to be more strict.
Added a spreadsheet with a truth table for the new tests.