Added abilty to xdg-open the locs-list for you
Added xdgOpenLocsListDialog that does xdg-open $LOCSLIST, to save the user having to type out the full path of the locs list file.
At the moment it's called only when the locs-list file is EMPTY.
This will help guide new users towards using the file. It's good if they're guided to it without having to be reading the text closely!