scanLocsList || xdgOpenLocsListDialogAndExit
xdgOpenLocsListDialog renamed to xdgOpenLocsListDialogAndExit, and now always exits the program after doing the dialog.
xdgOpenLocsListDialogAndExit is called if scanLocsList exits with an error value, i.e. if the scan finds any problems.
When the scan finds problems, this is the other major time when it's important to help the user get to the locs list quickly and painlessly.
However, the "scanLocsList || xdgOpenLocsListDialog" structure meant I needed to make changes to
ensure that the program EXITS if the scan finds errors.
I made the xdgOpenLocsListDialog do the exit, since the program should never be running while the user is editing the file!
In future I should consider if this exit is still a good idea even in the case where the user declines to edit the file...