Made the tester show holdall's output live
For the tester to work, I need the otuput of holdall to be saved in a variable.
It now also displays the program's output live (using tee).
Previously, it didn't show the output live.
This helps me by letting me use interactive mode with the tester - it lets me look at the dialogs, however, it invalidates the tests