Standardised use of getPermission, wrote dialogs
Now getPermission is used in a standardised way before every synchronise/merge command. Previously calls to synchronise/merge that were from a dialog did not getPermission because they were already a dialog. I thought it's clearer to the user if it works in a standard way.
Wrote more of the hostMissingDialog and deleteItem