deprecate writeToStatus and copyExitVal
writeToStatus is now split into several smaller functions, which
are called from the core logic section. This gives greater flexibility
to allow the status write to reflect the core logic's understanding
of the situation.
copyExitVal is now redundant, because in SAFE MODE a false exit value
aborts the program.
Since an error exit for the copy should be very unusual, and indicate
a serious problem, I think it is OK to panic exit instead of handle.